refactor: move JSONPreview to dedicated module
